Hard-code service name for Datadog
We use a hierarchy of tags on Datadog to provide different means for
filtering metrics and logs. All logs from a particular piece of
infrastructure share the same `service` tag and then have different tags
for the `env`. In this specific case, we want to tag everything as
coming from the `static.rust-lang.org` service and then differentiate
between the environments using the `dev` and `prod` valus for the
environment.